Fold in changes submitted by the PHP developers. (CVS 1742)
FossilOrigin-Name: a4a2570822f13ece2174438ef8f8a6175aaa4a6b

+++ src/table.c
@@ -31,7 +31,7 @@ typedef struct TabResult {
   int nAlloc;
   int nRow;
   int nColumn;
-  int nData;
+  long nData;
   int rc;
 } TabResult;
 
@@ -196,7 +196,7 @@ void sqlite_free_table(
     int i, n;
     azResult--;
     if( azResult==0 ) return;
-    n = (int)azResult[0];
+    n = (int)(long)azResult[0];
     for(i=1; i<n; i++){ if( azResult[i] ) free(azResult[i]); }
     free(azResult);
   }
